leveldb: move iStorage to leveldb
