+++ leveldb/db.go
@@ -168,7 +168,7 @@ func openDB(s *session) (*DB, error) {
 // The returned DB instance is safe for concurrent use.
 // The DB must be closed after use, by calling Close method.
 func Open(stor storage.Storage, o *opt.Options) (db *DB, err error) {
-	s, err := newSession(storage.IOCounterWrapper(stor), o)
+	s, err := newSession(iStorageWrapper(stor), o)
 	if err != nil {
 		return
 	}
@@ -963,7 +963,7 @@ func (db *DB) GetProperty(name string) (value string, err error) {
 		}
 	case p == "iostats":
 		var r, w float64
-		if s, ok := db.s.stor.(storage.IOCounter); ok {
+		if s, ok := db.s.stor.(IStorage); ok {
 			r, w = float64(s.Reads())/1048576.0, float64(s.Writes())/1048576.0
 		}
 		value = fmt.Sprintf("Read(MB): %13.5f Write(MB): %13.5f", r, w)
