+++ manualtest/dbstress/main.go
@@ -245,14 +245,14 @@ func (ts *testingStorage) Remove(fd storage.FileDesc) error {
 }
 
 func (ts *testingStorage) Reads() uint64 {
-	if s, ok := ts.Storage.(storage.IOCounter); ok {
+	if s, ok := ts.Storage.(leveldb.IStorage); ok {
 		return s.Reads()
 	}
 	return 0
 }
 
 func (ts *testingStorage) Writes() uint64 {
-	if s, ok := ts.Storage.(storage.IOCounter); ok {
+	if s, ok := ts.Storage.(leveldb.IStorage); ok {
 		return s.Writes()
 	}
 	return 0
